refactor: 테마 분리 및 MyApp 클래스 구조화
