+++ lib/app/my_app.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:device_preview_plus/device_preview_plus.dart';
+import 'package:baseball_diary/router.dart';
+import 'package:baseball_diary/theme/app_theme.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp.router(
+      routerConfig: router,
+      useInheritedMediaQuery: true,
+      locale: DevicePreview.locale(context),
+      builder: DevicePreview.appBuilder,
+      title: 'Flutter Demo',
+      themeMode: ThemeMode.system,
+      theme: AppTheme.lightTheme,
+      darkTheme: AppTheme.darkTheme,
+    );
+  }
+}
