+++ lib/main.dart
@@ -1,138 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:device_preview_plus/device_preview_plus.dart';
 import 'package:flutter/foundation.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:baseball_diary/router.dart';
+import 'package:baseball_diary/app/my_app.dart';
 
 void main() => runApp(
-  DevicePreview(
-    enabled: !kReleaseMode,
-    builder: (context) => const MyApp(), // 여기에서 MyApp을 실행
-  ),
+  DevicePreview(enabled: !kReleaseMode, builder: (context) => const MyApp()),
 );
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp.router(
-      routerConfig: router,
-      useInheritedMediaQuery: true,
-      locale: DevicePreview.locale(context),
-      builder: DevicePreview.appBuilder,
-      title: 'Flutter Demo',
-      themeMode: ThemeMode.system, // 시스템 다크모드에 따라 자동 적용
-      theme: _lightTheme(),
-      darkTheme: _darkTheme(),
-    );
-  }
-
-  ThemeData _lightTheme() {
-    return ThemeData(
-      brightness: Brightness.light,
-      textTheme: _textTheme(),
-      appBarTheme: const AppBarTheme(
-        backgroundColor: Colors.white,
-        foregroundColor: Colors.black,
-        elevation: 0,
-        titleTextStyle: TextStyle(
-          color: Colors.black,
-          fontSize: 18,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      scaffoldBackgroundColor: Colors.white,
-      primaryColor: const Color(0xFF1E2022),
-      bottomAppBarTheme: const BottomAppBarTheme(color: Colors.white),
-    );
-  }
-
-  ThemeData _darkTheme() {
-    return ThemeData(
-      brightness: Brightness.dark,
-      textTheme: _textTheme(),
-      appBarTheme: AppBarTheme(
-        backgroundColor: Colors.grey.shade900,
-        foregroundColor: Colors.white,
-        elevation: 0,
-        titleTextStyle: const TextStyle(
-          color: Colors.white,
-          fontSize: 18,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      scaffoldBackgroundColor: Colors.black,
-      primaryColor: const Color(0xFF52616A),
-      bottomAppBarTheme: BottomAppBarTheme(color: Colors.grey.shade900),
-    );
-  }
-
-  TextTheme _textTheme() {
-    return TextTheme(
-      displayLarge: GoogleFonts.gowunDodum(
-        fontSize: 96,
-        fontWeight: FontWeight.w300,
-        letterSpacing: -1.5,
-      ),
-      displayMedium: GoogleFonts.gowunDodum(
-        fontSize: 60,
-        fontWeight: FontWeight.w300,
-        letterSpacing: -0.5,
-      ),
-      displaySmall: GoogleFonts.gowunDodum(
-        fontSize: 48,
-        fontWeight: FontWeight.w500,
-      ),
-      headlineMedium: GoogleFonts.gowunDodum(
-        fontSize: 34,
-        fontWeight: FontWeight.w500,
-        letterSpacing: 0.25,
-      ),
-      headlineSmall: GoogleFonts.gowunDodum(
-        fontSize: 24,
-        fontWeight: FontWeight.w600,
-      ),
-      titleLarge: GoogleFonts.gowunDodum(
-        fontSize: 20,
-        fontWeight: FontWeight.w500,
-        letterSpacing: 0.15,
-      ),
-      titleMedium: GoogleFonts.gowunDodum(
-        fontSize: 16,
-        fontWeight: FontWeight.w600,
-        letterSpacing: 0.15,
-      ),
-      titleSmall: GoogleFonts.gowunDodum(
-        fontSize: 14,
-        fontWeight: FontWeight.w500,
-        letterSpacing: 0.1,
-      ),
-      bodyLarge: GoogleFonts.gowunDodum(
-        fontSize: 16,
-        fontWeight: FontWeight.w400,
-        letterSpacing: 0.5,
-      ),
-      bodyMedium: GoogleFonts.gowunDodum(
-        fontSize: 14,
-        fontWeight: FontWeight.w400,
-        letterSpacing: 0.25,
-      ),
-      labelLarge: GoogleFonts.gowunDodum(
-        fontSize: 14,
-        fontWeight: FontWeight.w500,
-        letterSpacing: 1.25,
-      ),
-      bodySmall: GoogleFonts.gowunDodum(
-        fontSize: 12,
-        fontWeight: FontWeight.w400,
-        letterSpacing: 0.4,
-      ),
-      labelSmall: GoogleFonts.gowunDodum(
-        fontSize: 10,
-        fontWeight: FontWeight.w400,
-        letterSpacing: 1.5,
-      ),
-    );
-  }
-}
