+++ lib/theme/app_theme.dart
@@ -1,0 +1,108 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class AppTheme {
+  static ThemeData lightTheme = ThemeData(
+    brightness: Brightness.light,
+    textTheme: _textTheme(),
+    appBarTheme: const AppBarTheme(
+      backgroundColor: Colors.white,
+      foregroundColor: Colors.black,
+      elevation: 0,
+      titleTextStyle: TextStyle(
+        color: Colors.black,
+        fontSize: 18,
+        fontWeight: FontWeight.bold,
+      ),
+    ),
+    scaffoldBackgroundColor: Colors.white,
+    primaryColor: const Color(0xFF1E2022),
+    bottomAppBarTheme: const BottomAppBarTheme(color: Colors.white),
+  );
+
+  static ThemeData darkTheme = ThemeData(
+    brightness: Brightness.dark,
+    textTheme: _textTheme(),
+    appBarTheme: AppBarTheme(
+      backgroundColor: Colors.grey.shade900,
+      foregroundColor: Colors.white,
+      elevation: 0,
+      titleTextStyle: const TextStyle(
+        color: Colors.white,
+        fontSize: 18,
+        fontWeight: FontWeight.bold,
+      ),
+    ),
+    scaffoldBackgroundColor: Colors.black,
+    primaryColor: const Color(0xFF52616A),
+    bottomAppBarTheme: BottomAppBarTheme(color: Colors.grey.shade900),
+  );
+
+  static TextTheme _textTheme() {
+    return TextTheme(
+      displayLarge: GoogleFonts.gowunDodum(
+        fontSize: 96,
+        fontWeight: FontWeight.w300,
+        letterSpacing: -1.5,
+      ),
+      displayMedium: GoogleFonts.gowunDodum(
+        fontSize: 60,
+        fontWeight: FontWeight.w300,
+        letterSpacing: -0.5,
+      ),
+      displaySmall: GoogleFonts.gowunDodum(
+        fontSize: 48,
+        fontWeight: FontWeight.w500,
+      ),
+      headlineMedium: GoogleFonts.gowunDodum(
+        fontSize: 34,
+        fontWeight: FontWeight.w500,
+        letterSpacing: 0.25,
+      ),
+      headlineSmall: GoogleFonts.gowunDodum(
+        fontSize: 24,
+        fontWeight: FontWeight.w600,
+      ),
+      titleLarge: GoogleFonts.gowunDodum(
+        fontSize: 20,
+        fontWeight: FontWeight.w500,
+        letterSpacing: 0.15,
+      ),
+      titleMedium: GoogleFonts.gowunDodum(
+        fontSize: 16,
+        fontWeight: FontWeight.w600,
+        letterSpacing: 0.15,
+      ),
+      titleSmall: GoogleFonts.gowunDodum(
+        fontSize: 14,
+        fontWeight: FontWeight.w500,
+        letterSpacing: 0.1,
+      ),
+      bodyLarge: GoogleFonts.gowunDodum(
+        fontSize: 16,
+        fontWeight: FontWeight.w400,
+        letterSpacing: 0.5,
+      ),
+      bodyMedium: GoogleFonts.gowunDodum(
+        fontSize: 14,
+        fontWeight: FontWeight.w400,
+        letterSpacing: 0.25,
+      ),
+      labelLarge: GoogleFonts.gowunDodum(
+        fontSize: 14,
+        fontWeight: FontWeight.w500,
+        letterSpacing: 1.25,
+      ),
+      bodySmall: GoogleFonts.gowunDodum(
+        fontSize: 12,
+        fontWeight: FontWeight.w400,
+        letterSpacing: 0.4,
+      ),
+      labelSmall: GoogleFonts.gowunDodum(
+        fontSize: 10,
+        fontWeight: FontWeight.w400,
+        letterSpacing: 1.5,
+      ),
+    );
+  }
+}
